Clean up the huge portability regressions caused by the introduction of ZPL and dependent features.
This massively guts zpl.h: everything we don't use. It also improves the portability of the things we do use.
See zpl.h for why keeping the entire library is unacceptable.

Some notes:
Intel, or more specifically x86, is an architecture, not a platform.
The uses of zpl_vm_alloc functions contained a regression in that unlike the X-prefixed allocation functions, no checking of the return value was performed. Moreover, the new Bgetpagesize incorrectly used the zpl_virtual_memory_page_size API, causing the value returned to be dwAllocationGranularity instead of dwPageSize on Windows!
timerInit contained a compilation error on "MIPS".
The introduction of sysReadCPUID() broke linking on everything but x86.
Any symbol containing two successive underscores is reserved, so don't use that for include guards.

git-svn-id: https://svn.eduke32.com/eduke32@8273 1a8010ca-5511-0410-912e-c29ae57300e0