Replace zpl_yield_thread in DirectSound driver
The PAUSE instruction it issues on x86 tells hyperthreading processors to deprioritize the current instruction stream and is often used for spinlocks. The Apple implementation is erroneous: a compiler-level memory fence that compiles to nothing.

Replace it with SwitchToYield, which tells the OS to context switch, as intended.

From-SVN: r8277