Update the Microsoft resource compiler scripts used for Windows binaries
This adds the version control revision plus build time to the properties of the resulting binary, and updates some of the legal text. The executable version number has also been bumped from the longstanding "1.9.9.9" to "2.0.0.4", symbolic of the year EDuke32 was first released in. Future Fury binaries will now say "2.0.1.9" to match its first non-preview release as well.

git-svn-id: https://svn.eduke32.com/eduke32@8365 1a8010ca-5511-0410-912e-c29ae57300e0

# Conflicts:
#	Common.mak