PCExhumed: Further improve the widescreen fix for the smoke in menu_DrawTheMap. rotatesprite is now directly called, instead of going through the overwritesprite wrapper, in order to trim the smoke sprites, so they don't leave the map's boundaries in widescreen.
A technical side-effect is that picanm[tilenum].sf is not temporarily
set to 0, as done in overwritesprite.

The call to videoClearScreen is still required, in case the console is
shown and then hidden. It's now called after HandleAsync, just for the
sake of it (less useful in case of quit, and maybe also in another scenario).