Duke3d: aim assist for game controllers
A helpful side-effect of this commit is that several of the previous changes to how screen re-centering etc work can now be deterministic again, which somewhat fixes demo recording and playback. For now, recording a demo will limit input to the game tickrate like it did in the old days.