Duke3d: fix hiccup in movement at internal CAMERA1 panning transition points
The way this code works is pretty weird. It seems like the movement hiccup here was caused by intentionally skippping the state of the camera where pData[1] == 0, but I can't find anything else in the code that depends on the value of this field for CAMERA1, and fixing the hiccup causes no issues in any of the setups with panning cameras that I looked at for testing.