engine: tweak/"optimize" the C version of nsqrtasm()
Disassembly of the tweaked version versus the original is very similar, but MSVC doesn't seem to want to inline the original version without forcing it. I tested the performance of the change under GCC using a very large loop and the results were in favor of the tweaked version by a negligible margin. FORCE_INLINE was added after the fact.