Duke3D: Re-enable remote activation of viewscreen using yvel of 1
A viewscreen placed in the map with a yvel of 1 will always display
its camera feed if the player is within active range. If set to 0,
the screen will only show the tile on manual activation.

If yvel is 1, and the screen is connected to a camera, it will display
the view of that camera. Otherwise, it will show the view of the
Viewscreen's owner, the direction in which it is looking being
dictated by angle and shade of that sprite.

The xvel of the viewscreen now defines the active range.
If less or equal to 0, the default distance of 8192 is used.

Note that this is currently restricted to one active screen at a time.

Fixes terminx/eduke32#41