Duke3d: add display selection to menu, tweak menu framerate limit option
I'm not very happy with the list of values for the framerate limiter. All methods of presenting these values to the end user seem to be terrible: linear slider bars are cumbersome and contain many unnecessary values, and simply presenting a field for a value to be entered into requires the end user know what they're doing. Ugh.