engine: polymer: 'fix' texture scale on relative floors
Classic relative scaling apparently had some numerical artifacts related
to length of firstwall. Cribbing the logic added in 6fc6b7d5 in Polymost
along with bumping the precision of the intermediate terms seems to make
it line up fully. Rooftop floor in E1L1 and surrounding ledges are good
test cases for this.