Duke3D: Add actor struct members `htfloorzoffset` and `htwaterzoffset`
Values range from -128 to 127. The former defines the distance of the
actor from all floors, the latter applies additively to the hardcoded
offset applied when on lotag 1 sectors. Positive values lift the actor
off the ground, negative values move the actor into the floor.

The default values for both struct members is 0. To achieve a net zero
displacement, htfloorzoffset needs to be set to -1, and htwaterzoffset
must be set to 24. Also note that internally, the offset is multiplied
by 256 to receive the actual z coordinates (like `iffloordistl`).