Duke3D: Only attempt to switch subweapon if we actually have the weapon
This fixes a bug in World Tour where trying to switch to the flamethrower
while the current weapon is the freezer, and you have flamethrower ammo
but not the actual weapon itself, then this instead locks the weapon key,
preventing you from switching back to the freezer from other weapons.

This bug also affected the Shrinker, but only if CON code changes the
'gotweapon 11' value to 0 after the player obtains the weapon.