Duke3d: fix .ogg sounds defined for Duke talk sometimes not playing when used in the skill selection menu
We were hitting the FX_SoundActive() check for ud.skill_voice in G_NewGame() before the worker threads completed the decoder initialization work scheduled by the S_PlaySound() call in the MENU_SKILL case in Menu_EntryLinkActivate()--the code in G_NewGame() was proceeding to the next stage because FX_SoundActive() returned false, but our check occurred before it had a chance to return true.

This also swaps the call to FX_StopAllSounds() in G_EnterLevel() out for a call to S_StopAllSounds() instead, and alters S_StopAllSounds() to set g_dukeTalk to false.

Fixes terminx/eduke32#215