Duke3D: Introduce actor event 'EVENT_PREACTORDAMAGE'
This event is executed when an actor runs A_IncurDamage(), and is
placed just before subtracting 'htextra' from 'extra' and updating
the actor's owner (i.e. just before damage is applied to the actor).

For useractors, this normally occurs when ifhitweapon is called.

Set RETURN to a nonzero value to skip applying the damage. This
will also cause the 'ifhitweapon' conditional to skip the branch.
Note that doing so will not automatically reset htextra to 0.