engine: enable large page support in mimalloc
I think changing the default value here is the only way to ensure the option is set at process initialization time, before control is passed to main(). It looks like the "official" way to enable this is to set some environmental variables before spawning the process, but I don't think that solution is really appropriate for production builds.