Blood WLB: Changes from What Lies Beneath source
The changes that BloodyTom did to NBlood for his What Lies Beneath
mod. He used commit "c8421de2 Fix rumble detection (#823)" as a base,
I rebased this onto the current master code.
Otherwise this is almost unaltered, except I uncommented an assertion
to properly fix it in a later commit