Blood: Write savegames to temp file first, rename when done
If the game crashed while saving, the savegame was corrupted.
Prevent that by saving to a temporary file first and renaming that
to the proper savegame name when saving is done (without crashing),
so at least the previous version of the savegame is preserved in case
of crashes.
Additionally this creates a backup of the previous version as
gameXXX.sav_bk in case I regret doing a quicksave